pyproject.toml
[tool.poetry]
name = "llm-graph-search"
version = "0.1.0"
description = ""
authors = ["riaz <riaz.2012@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
neo4j = "^5.21.0"
gremlinpython = "^3.7.2"
openai = "^1.35.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
